Extract code to verify VP token response
